Refactor Home component to improve code health by extracting RefillReminders and RecentOrders into separate components. This reduces the file size and complexity of Home.tsx.
